Rename enum store dictionary api and implementation and move to separate files.
