+++ searchlib/src/vespa/searchlib/attribute/enumstorebase.cpp
@@ -2,6 +2,7 @@
 
 #include "enumstorebase.h"
 #include "enumstore.h"
+#include "enum_store_dictionary.h"
 #include <vespa/vespalib/btree/btree.hpp>
 #include <vespa/vespalib/btree/btreeiterator.hpp>
 #include <vespa/vespalib/btree/btreenode.hpp>
@@ -74,9 +75,9 @@ EnumStoreBase::EnumStoreBase(uint64_t initBufferSize, bool hasPostings)
       _toHoldBuffers()
 {
     if (hasPostings)
-        _enumDict = new EnumStoreDict<EnumPostingTree>(*this);
+        _enumDict = new EnumStoreDictionary<EnumPostingTree>(*this);
     else
-        _enumDict = new EnumStoreDict<EnumTree>(*this);
+        _enumDict = new EnumStoreDictionary<EnumTree>(*this);
     _store.addType(&_type);
     _type.setSizeNeededAndDead(initBufferSize, 0);
     _store.initActiveBuffers();
@@ -249,220 +250,7 @@ vespalib::asciistream & operator << (vespalib::asciistream & os, const EnumStore
 }
 
 
-EnumStoreDictBase::EnumStoreDictBase()
-    : datastore::UniqueStoreDictionaryBase()
-{
-}
-
-EnumStoreDictBase::~EnumStoreDictBase() = default;
-
-template <typename Dictionary>
-EnumStoreDict<Dictionary>::EnumStoreDict(EnumStoreBase &enumStore)
-    : ParentUniqueStoreDictionary(),
-      _enumStore(enumStore)
-{
-}
-
-template <typename Dictionary>
-EnumStoreDict<Dictionary>::~EnumStoreDict() = default;
-
-template <typename Dictionary>
-uint32_t
-EnumStoreDict<Dictionary>::getNumUniques() const
-{
-    return this->_dict.size();
-}
-
-template <typename Dictionary>
-void
-EnumStoreDict<Dictionary>::
-writeAllValues(BufferWriter &writer,
-               btree::BTreeNode::Ref rootRef) const
-{
-    constexpr size_t BATCHSIZE = 1000;
-    std::vector<Index> idxs;
-    idxs.reserve(BATCHSIZE);
-    typename Dictionary::Iterator it(rootRef, this->_dict.getAllocator());
-    while (it.valid()) {
-        if (idxs.size() >= idxs.capacity()) {
-            _enumStore.writeValues(writer, &idxs[0], idxs.size());
-            idxs.clear();
-        }
-        idxs.push_back(it.getKey());
-        ++it;
-    }
-    if (!idxs.empty()) {
-        _enumStore.writeValues(writer, &idxs[0], idxs.size());
-    }
-}
-
-template <typename Dictionary>
-ssize_t
-EnumStoreDict<Dictionary>::deserialize(const void *src,
-                                       size_t available,
-                                       IndexVector &idx)
-{
-    return _enumStore.deserialize(src, available, idx, this->_dict);
-}
-
-template <typename Dictionary>
-void
-EnumStoreDict<Dictionary>::fixupRefCounts(const EnumVector & hist)
-{
-    _enumStore.fixupRefCounts(hist, this->_dict);
-}
-
-template <typename Dictionary>
-void
-EnumStoreDict<Dictionary>::removeUnusedEnums(const IndexSet &unused,
-                                             const datastore::EntryComparator &cmp,
-                                             const datastore::EntryComparator *fcmp)
-{
-    typedef typename Dictionary::Iterator Iterator;
-    if (unused.empty())
-        return;
-    Iterator it(BTreeNode::Ref(), this->_dict.getAllocator());
-    for (const auto& idx : unused) {
-        it.lower_bound(this->_dict.getRoot(), idx, cmp);
-        assert(it.valid() && !cmp(idx, it.getKey()));
-        if (Iterator::hasData() && fcmp != nullptr) {
-            typename Dictionary::DataType pidx(it.getData());
-            this->_dict.remove(it);
-            if (!it.valid() || (*fcmp)(idx, it.getKey())) {
-                continue;  // Next entry does not use same posting list
-            }
-            --it;
-            if (it.valid() && !(*fcmp)(it.getKey(), idx)) {
-                continue;  // Previous entry uses same posting list
-            }
-            if (it.valid()) {
-                ++it;
-            } else {
-                it.begin();
-            }
-            this->_dict.thaw(it);
-            it.writeData(pidx);
-        } else {
-            this->_dict.remove(it);
-        }
-   }
-}
-
-template <typename Dictionary>
-void
-EnumStoreDict<Dictionary>::freeUnusedEnums(const datastore::EntryComparator &cmp,
-                                           const datastore::EntryComparator *fcmp)
-{
-    IndexSet unused;
-
-    // find unused enums
-    for (auto iter = this->_dict.begin(); iter.valid(); ++iter) {
-        _enumStore.freeUnusedEnum(iter.getKey(), unused);
-    }
-    removeUnusedEnums(unused, cmp, fcmp);
-}
-
-template <typename Dictionary>
-void
-EnumStoreDict<Dictionary>::freeUnusedEnums(const IndexSet& toRemove,
-                                           const datastore::EntryComparator& cmp,
-                                           const datastore::EntryComparator* fcmp)
-{
-    IndexSet unused;
-    for (const auto& index : toRemove) {
-        _enumStore.freeUnusedEnum(index, unused);
-    }
-    removeUnusedEnums(unused, cmp, fcmp);
-}
-
-template <typename Dictionary>
-bool
-EnumStoreDict<Dictionary>::findIndex(const datastore::EntryComparator &cmp,
-                                     Index &idx) const
-{
-    auto itr = this->_dict.find(Index(), cmp);
-    if (!itr.valid()) {
-        return false;
-    }
-    idx = itr.getKey();
-    return true;
-}
-
-template <typename Dictionary>
-bool
-EnumStoreDict<Dictionary>::findFrozenIndex(const datastore::EntryComparator &cmp,
-                                           Index &idx) const
-{
-    auto itr = this->_dict.getFrozenView().find(Index(), cmp);
-    if (!itr.valid()) {
-        return false;
-    }
-    idx = itr.getKey();
-    return true;
-}
-
-template <typename Dictionary>
-std::vector<EnumStoreBase::EnumHandle>
-EnumStoreDict<Dictionary>::findMatchingEnums(const datastore::EntryComparator &cmp) const
-{
-    std::vector<EnumStoreBase::EnumHandle> result;
-    auto itr = this->_dict.getFrozenView().find(Index(), cmp);
-    while (itr.valid() && !cmp(Index(), itr.getKey())) {
-        result.push_back(itr.getKey().ref());
-        ++itr;
-    }
-    return result;
-}
-
-template <typename Dictionary>
-void
-EnumStoreDict<Dictionary>::onReset()
-{
-    this->_dict.clear();
-}
-
-
-template <>
-EnumPostingTree &
-EnumStoreDict<EnumTree>::getPostingDictionary()
-{
-    LOG_ABORT("should not be reached");
-}
-
-
-template <>
-EnumPostingTree &
-EnumStoreDict<EnumPostingTree>::getPostingDictionary()
-{
-    return _dict;
-}
-
-
-template <>
-const EnumPostingTree &
-EnumStoreDict<EnumTree>::getPostingDictionary() const
-{
-    LOG_ABORT("should not be reached");
-}
-
-
-template <>
-const EnumPostingTree &
-EnumStoreDict<EnumPostingTree>::getPostingDictionary() const
-{
-    return _dict;
-}
-
-
-template <typename Dictionary>
-bool
-EnumStoreDict<Dictionary>::hasData() const
-{
-    return Dictionary::LeafNodeType::hasData();
-}
-
-
-template class datastore::DataStoreT<datastore::AlignedEntryRefT<31, 4> >;
+template class datastore::DataStoreT<EnumStoreIndex>;
 
 template
 ssize_t
@@ -479,103 +267,6 @@ EnumStoreBase::fixupRefCounts<EnumTree>(const EnumVector &hist, EnumTree &tree);
 template
 void
 EnumStoreBase::fixupRefCounts<EnumPostingTree>(const EnumVector &hist, EnumPostingTree &tree);
-
-template class EnumStoreDict<EnumTree>;
-
-template class EnumStoreDict<EnumPostingTree>;
-
-template
-class btree::BTreeNodeT<EnumStoreBase::Index, EnumTreeTraits::INTERNAL_SLOTS>;
-
-template
-class btree::BTreeNodeTT<EnumStoreBase::Index, datastore::EntryRef, btree::NoAggregated, EnumTreeTraits::INTERNAL_SLOTS>;
-
-template
-class btree::BTreeNodeTT<EnumStoreBase::Index, btree::BTreeNoLeafData, btree::NoAggregated, EnumTreeTraits::LEAF_SLOTS>;
-
-template
-class btree::BTreeInternalNode<EnumStoreBase::Index, btree::NoAggregated, EnumTreeTraits::INTERNAL_SLOTS>;
-
-template
-class btree::BTreeLeafNode<EnumStoreBase::Index, btree::BTreeNoLeafData, btree::NoAggregated, EnumTreeTraits::LEAF_SLOTS>;
-
-template
-class btree::BTreeLeafNode<EnumStoreBase::Index, datastore::EntryRef, btree::NoAggregated, EnumTreeTraits::LEAF_SLOTS>;
-
-template
-class btree::BTreeLeafNodeTemp<EnumStoreBase::Index, btree::BTreeNoLeafData, btree::NoAggregated, EnumTreeTraits::LEAF_SLOTS>;
-
-template
-class btree::BTreeLeafNodeTemp<EnumStoreBase::Index, datastore::EntryRef, btree::NoAggregated, EnumTreeTraits::LEAF_SLOTS>;
-
-template
-class btree::BTreeNodeStore<EnumStoreBase::Index, btree::BTreeNoLeafData, btree::NoAggregated,
-                            EnumTreeTraits::INTERNAL_SLOTS, EnumTreeTraits::LEAF_SLOTS>;
-
-template
-class btree::BTreeNodeStore<EnumStoreBase::Index, datastore::EntryRef, btree::NoAggregated,
-                            EnumTreeTraits::INTERNAL_SLOTS, EnumTreeTraits::LEAF_SLOTS>;
-
-
-template
-class btree::BTreeRoot<EnumStoreBase::Index, btree::BTreeNoLeafData, btree::NoAggregated,
-                       const datastore::EntryComparatorWrapper, EnumTreeTraits>;
-
-template
-class btree::BTreeRoot<EnumStoreBase::Index, datastore::EntryRef, btree::NoAggregated,
-                       const datastore::EntryComparatorWrapper, EnumTreeTraits>;
-
-template
-class btree::BTreeRootT<EnumStoreBase::Index, btree::BTreeNoLeafData, btree::NoAggregated,
-                        const datastore::EntryComparatorWrapper, EnumTreeTraits>;
-
-template
-class btree::BTreeRootT<EnumStoreBase::Index, datastore::EntryRef, btree::NoAggregated,
-                        const datastore::EntryComparatorWrapper, EnumTreeTraits>;
-
-template
-class btree::BTreeRootBase<EnumStoreBase::Index, btree::BTreeNoLeafData, btree::NoAggregated,
-                           EnumTreeTraits::INTERNAL_SLOTS, EnumTreeTraits::LEAF_SLOTS>;
-
-template
-class btree::BTreeRootBase<EnumStoreBase::Index, datastore::EntryRef, btree::NoAggregated,
-                           EnumTreeTraits::INTERNAL_SLOTS, EnumTreeTraits::LEAF_SLOTS>;
-
-template
-class btree::BTreeNodeAllocator<EnumStoreBase::Index, btree::BTreeNoLeafData, btree::NoAggregated,
-                                EnumTreeTraits::INTERNAL_SLOTS, EnumTreeTraits::LEAF_SLOTS>;
-
-template
-class btree::BTreeNodeAllocator<EnumStoreBase::Index, datastore::EntryRef, btree::NoAggregated,
-                                EnumTreeTraits::INTERNAL_SLOTS, EnumTreeTraits::LEAF_SLOTS>;
-
-template
-class btree::BTreeIteratorBase<EnumStoreBase::Index, btree::BTreeNoLeafData, btree::NoAggregated,
-                               EnumTreeTraits::INTERNAL_SLOTS, EnumTreeTraits::LEAF_SLOTS, EnumTreeTraits::PATH_SIZE>;
-template
-class btree::BTreeIteratorBase<EnumStoreBase::Index, datastore::EntryRef, btree::NoAggregated,
-                               EnumTreeTraits::INTERNAL_SLOTS, EnumTreeTraits::LEAF_SLOTS, EnumTreeTraits::PATH_SIZE>;
-
-template class btree::BTreeConstIterator<EnumStoreBase::Index, btree::BTreeNoLeafData, btree::NoAggregated,
-                                         const datastore::EntryComparatorWrapper, EnumTreeTraits>;
-
-template class btree::BTreeConstIterator<EnumStoreBase::Index, datastore::EntryRef, btree::NoAggregated,
-                                         const datastore::EntryComparatorWrapper, EnumTreeTraits>;
-
-template
-class btree::BTreeIterator<EnumStoreBase::Index, btree::BTreeNoLeafData, btree::NoAggregated,
-                           const datastore::EntryComparatorWrapper, EnumTreeTraits>;
-template
-class btree::BTreeIterator<EnumStoreBase::Index, datastore::EntryRef, btree::NoAggregated,
-                           const datastore::EntryComparatorWrapper, EnumTreeTraits>;
-
-template
-class btree::BTree<EnumStoreBase::Index, btree::BTreeNoLeafData, btree::NoAggregated,
-                   const datastore::EntryComparatorWrapper, EnumTreeTraits>;
-template
-class btree::BTree<EnumStoreBase::Index, datastore::EntryRef, btree::NoAggregated,
-                   const datastore::EntryComparatorWrapper, EnumTreeTraits>;
-
 
 }
